Reduce code duplication -- move `dual_step` to `ConstraintOptimizer`
